% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Queue2}
\alias{Queue2}
\title{A queue implemented entirely in R}
\usage{
Queue2(init = 20)
}
\arguments{
\item{init}{Initial size of the list that backs the queue. This is also used
as the minimum size of the list; it will not shrink any smaller.}
}
\description{
A Queue2 object is backed by a list, which is used in a circular manner. All
manipulation is done in R, without calling out to any compiled code. The
backing list will grow or shrink as the queue changes in size.
}

