% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Stack2}
\alias{Stack2}
\title{A stack implemented entirely in R}
\usage{
Stack2(init = 20)
}
\arguments{
\item{init}{Initial size of the list that backs the stack. This is also used
as the minimum size of the list; it will not shrink any smaller.}
}
\description{
A Stack2 object is backed by a list. All manipulation is done in R, without
calling out to any compiled code. The backing list will grow or shrink as the
stack changes in size.
}

